LPS-59277 - Register command for EditProductVersionAction action
