+++ portal-web/docroot/WEB-INF/tiles-defs.xml
@@ -209,10 +209,6 @@
 	<!-- Software Catalog -->
 	<definition name="portlet.software_catalog" extends="portlet" />
 
-	<definition name="portlet.software_catalog.edit_product_version" extends="portlet.software_catalog">
-		<put name="portlet_content" value="/portlet/software_catalog/edit_product_version.jsp" />
-	</definition>
-
 	<definition name="portlet.software_catalog.error" extends="portlet.software_catalog">
 		<put name="portlet_content" value="/portlet/software_catalog/error.jsp" />
 	</definition>
